fix: Eliminate need for Product.js and instead have components use new custom hook useProductState.js
Add to cart on ProductPage and remove from cart on ShoppingCart now behave as expected. User is no longer logged on a window refresh but instead the component rerenders on remove from cart.

All API calls to back end all seem to work as expected. <---!!! OK

Known issue: Add to favorites and remove to favorites seem to work as expected, except when the user adds to favorites then refreshes the page, add to favorites button is rendered and not remove from favorites. Also remove from favorites does not rerender on Favorites page.